Remove JobManager Class
Move outputStatistics to RunMachineShop
and setupJobs to MachineShopSimulator, leaving
us able to delete JobManager as per reviewer
recommendation on our master pull request.
